Update README.md package naming on receivers
This might be easier for new users to the app if rather than relying on the user spotting the placeholders you have left. Instead if you use `${applicationId}` then the Gradle Manifest Merger will take care of the package name. This is particularly useful for people who are building multiple product flavours of an app.